Jackpieking-M7: Done applying CQRS pattern for skill entity, creating new Image.Firebase, Mail.GoogleGmail, Cache.Redis layer.

+++ src/Persistence.SqlServer2016/Repositories/BlogCommentRepository.cs
@@ -25,6 +25,11 @@ internal sealed class BlogCommentRepository :
         Guid authorId,
         CancellationToken cancellationToken)
     {
+        if (authorId == Guid.Empty)
+        {
+            return Task.FromResult<int>(result: default);
+        }
+
         return _dbSet
             .Where(predicate: blogComment => blogComment.AuthorId == authorId)
             .ExecuteDeleteAsync(cancellationToken: cancellationToken);
