+++ src/Persistence.SqlServer2016/Repositories/PlatformRepository.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Domain.Entities;
 using Domain.Repositories;
 using Microsoft.EntityFrameworkCore;
+using Persistence.SqlServer2016.Common;
 using Persistence.SqlServer2016.Data;
 using Persistence.SqlServer2016.Repositories.Base;
 
@@ -25,17 +26,30 @@ internal sealed class PlatformRepository :
         Guid platformId,
         CancellationToken cancellationToken)
     {
+        if (platformId == Guid.Empty)
+        {
+            return Task.FromResult<int>(result: default);
+        }
+
         return _dbSet
             .Where(predicate: platform => platform.Id == platformId)
             .ExecuteDeleteAsync(cancellationToken: cancellationToken);
     }
 
-    public Task<int> BulkUpdateByPlatformIdAsync(
+    public Task<int> BulkUpdateByPlatformIdVer1Async(
         Guid platformId,
         DateTime platformRemovedAt,
         Guid platformRemovedBy,
         CancellationToken cancellationToken)
     {
+        if (platformId == Guid.Empty ||
+            platformRemovedAt < CommonConstant.DbDefaultValue.MIN_DATE_TIME ||
+            platformRemovedAt > DateTime.UtcNow ||
+            platformRemovedBy == Guid.Empty)
+        {
+            return Task.FromResult<int>(result: default);
+        }
+
         return _dbSet
             .Where(predicate: platform => platform.Id == platformId)
             .ExecuteUpdateAsync(
@@ -49,11 +63,20 @@ internal sealed class PlatformRepository :
                 cancellationToken: cancellationToken);
     }
 
-    public Task<int> BulkUpdateByPlatformIdAsync(
+    public Task<int> BulkUpdateByPlatformIdVer2Async(
         Guid platformId,
         string platformName,
         CancellationToken cancellationToken)
     {
+        const int MaxPlatformNameLength = 100;
+
+        if (platformId == Guid.Empty ||
+            string.IsNullOrWhiteSpace(value: platformName) ||
+            platformName.Length > MaxPlatformNameLength)
+        {
+            return Task.FromResult<int>(result: default);
+        }
+
         return _dbSet
             .Where(predicate: platform => platform.Id == platformId)
             .ExecuteUpdateAsync(
