+++ src/Persistence.SqlServer2016/Repositories/RefreshTokenRepository.cs
@@ -25,6 +25,11 @@ internal sealed class RefreshTokenRepository :
         Guid userId,
         CancellationToken cancellationToken)
     {
+        if (userId == Guid.Empty)
+        {
+            return Task.FromResult<int>(result: default);
+        }
+
         return _dbSet
             .Where(refreshToken => refreshToken.UserId == userId)
             .ExecuteDeleteAsync(cancellationToken: cancellationToken);
